+++ c2c-extension-test/build.gradle
@@ -28,6 +28,7 @@ dependencies {
     implementation project(":c2c-extension")
     implementation project(":c2c-tooling")
 
+    testImplementation project(":c2c-util")
     testImplementation "org.ballerinalang:ballerina-lang:${ballerinaLangVersion}"
     testImplementation "com.github.docker-java:docker-java:${dockerJavaVersion}"
     testImplementation "io.fabric8:kubernetes-client:${fabric8KubernetesClientVersion}"
@@ -37,9 +38,7 @@ dependencies {
     testImplementation "org.ballerinalang:language-server-commons:${ballerinaLangVersion}"
     testImplementation "org.ballerinalang:language-server-cli:${ballerinaLangVersion}"
     testImplementation "org.ballerinalang:ballerina-cli:${ballerinaLangVersion}"
-
-
-
+    
     testImplementation 'org.eclipse.lsp4j:org.eclipse.lsp4j:0.12.0'
     testImplementation 'org.eclipse.lsp4j:org.eclipse.lsp4j.jsonrpc:0.12.0'
 
@@ -53,15 +52,6 @@ description = 'Ballerina - C2C Extension Tests'
 
 def distributionLocation = configurations.bDistributionDir.singleFile
 def zipDirLocation = configurations.zipPartent.singleFile
-task copyExtensionPlugin {
-    dependsOn configurations.extensionPlugin
-    doLast {
-        copy {
-            from configurations.extensionPlugin
-            into file("${distributionLocation}/bre/lib/")
-        }
-    }
-}
 
 task copyTooling {
     dependsOn configurations.tooling
@@ -80,16 +70,10 @@ task copyBallerinaArtifacts {
             from file("$zipDirLocation/bala/ballerina")
             into file("${distributionLocation}/repo/bala/ballerina")
         }
-        copy {
-            from file("$zipDirLocation/cache/ballerina")
-            into file("${distributionLocation}/repo/cache/ballerina")
-        }
-
     }
 }
 
 task prepareDistribution {
-    dependsOn copyExtensionPlugin
     dependsOn copyBallerinaArtifacts
     dependsOn copyTooling
 }
@@ -107,7 +91,6 @@ test {
                 "ballerina.version" : "${version}",
                 "experimental" : "true",
                 sampleDir      : System.getProperty('sampleDir', project.projectDir.absolutePath + "/../samples"),
-                ballerinaPack  : System.getProperty('ballerinaPack', "$configurations.bDistributionDir.singleFile.absolutePath"),
                 jacocoAgentLine: "-javaagent:${configurations.jacocoRuntime.asPath}=destfile=${test.jacoco.destinationFile.absolutePath},append=true,inclnolocationclasses=true,exclclassloader=jdk.internal.reflect.DelegatingClassLoader"
         ]
     }
