+++ c2c-extension-test/src/test/resources/testng.xml
@@ -45,8 +45,8 @@
             <class name="io.ballerina.c2c.test.JobTest"/>
             <class name="io.ballerina.c2c.test.EnvTest"/>
             <class name="io.ballerina.c2c.test.CustomDiagnosticsTest"/>
-<!--            <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidServicePath"/>-->
-<!--            <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidResourcePath"/>-->
+            <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidServicePath"/>
+            <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidResourcePath"/>
             <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidResourcePathEmpty"/>
             <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidResourcePathEquals"/>
             <class name="io.ballerina.c2c.test.codeactions.diagnostic.InvalidPort"/>
