--- c2c-extension/src/main/java/io/ballerina/c2c/diagnostics/HttpsConfig.java
@@ -1,46 +1,0 @@
-/*
- *  Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-package io.ballerina.c2c.diagnostics;
-
-import java.util.Optional;
-
-/**
- * Represents a Http config of a secure socket.
- *
- * @since 2.0.0
- */
-public class HttpsConfig {
-    private SecureSocketConfig secureSocketConfig;
-    private MutualSSLConfig mutualSSLConfig;
-
-    public Optional<SecureSocketConfig> getSecureSocketConfig() {
-        return Optional.ofNullable(secureSocketConfig);
-    }
-
-    public void setSecureSocketConfig(SecureSocketConfig secureSocketConfig) {
-        this.secureSocketConfig = secureSocketConfig;
-    }
-
-    public Optional<MutualSSLConfig> getMutualSSLConfig() {
-        return Optional.ofNullable(mutualSSLConfig);
-    }
-
-    public void setMutualSSLConfig(MutualSSLConfig mutualSSLConfig) {
-        this.mutualSSLConfig = mutualSSLConfig;
-    }
-}
