+++ c2c-extension/src/main/java/io/ballerina/c2c/handlers/DockerHandler.java
@@ -18,9 +18,9 @@
 
 package io.ballerina.c2c.handlers;
 
-import io.ballerina.c2c.diagnostics.C2CDiagnosticCodes;
 import io.ballerina.c2c.diagnostics.NullLocation;
 import io.ballerina.c2c.exceptions.KubernetesPluginException;
+import io.ballerina.c2c.util.C2CDiagnosticCodes;
 import io.ballerina.tools.diagnostics.Diagnostic;
 import io.ballerina.tools.diagnostics.DiagnosticFactory;
 import io.ballerina.tools.diagnostics.DiagnosticInfo;
