+++ c2c-util/src/main/java/io/ballerina/c2c/util/ProjectServiceInfo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.ballerina.c2c.diagnostics;
+package io.ballerina.c2c.util;
 
 import io.ballerina.compiler.api.SemanticModel;
 import io.ballerina.compiler.syntax.tree.Node;
