+++ c2c-util/src/main/java/io/ballerina/c2c/util/ResourceInfo.java
@@ -15,7 +15,7 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.c2c.diagnostics;
+package io.ballerina.c2c.util;
 
 import io.ballerina.compiler.syntax.tree.FunctionDefinitionNode;
