+++ c2c-util/src/main/java/io/ballerina/c2c/util/SecureSocketConfig.java
@@ -15,31 +15,40 @@
  *  specific language governing permissions and limitations
  *  under the License.
  */
-package io.ballerina.c2c.diagnostics;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.util.Optional;
+package io.ballerina.c2c.util;
 
 /**
- * Represents Listener information of a ballerina document.
+ * Represents a Config inside a secure socket.
  *
  * @since 2.0.0
  */
-@Getter
-@Setter
-public class ListenerInfo {
-    private String name;
-    private int port;
-    private HttpsConfig config;
+public class SecureSocketConfig {
+    private String certFile;
+    private String keyFile;
+    private String path;
 
-    public ListenerInfo(String name, int port) {
-        this.name = name;
-        this.port = port;
+
+    public String getCertFile() {
+        return certFile;
     }
 
-    public Optional<HttpsConfig> getConfig() {
-        return Optional.ofNullable(config);
+    public String getKeyFile() {
+        return keyFile;
+    }
+
+    public String getPath() {
+        return path;
+    }
+
+    public void setPath(String path) {
+        this.path = path;
+    }
+
+    public void setKeyFile(String keyFile) {
+        this.keyFile = keyFile;
+    }
+
+    public void setCertFile(String certFile) {
+        this.certFile = certFile;
     }
 }
