+++ settings.gradle
@@ -25,6 +25,7 @@ include(':c2c-extension')
 include(':c2c-ballerina')
 include(':c2c-extension-test')
 include (':c2c-tooling')
+include (':c2c-util')
 
 gradleEnterprise {
     buildScan {
